+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Block selection appending duplicate newlines when last column is selected
 - Bitmap fonts being a bit smaller than they should be in some cases
 - Config reload creating alternate screen history instead of updating scrollback
+- Ongoing selection not updating when content moves below it
 
 ### Removed